+++ rst2pdf/basenodehandler.py
@@ -36,10 +36,10 @@ will be logged.
 import inspect
 import types
 
+from smartypants import smartypants
 import docutils.nodes
 
 from .flowables import BoundByWidth, TocEntry
-from .smartypants import smartyPants
 from .log import log, nodeid
 
 
@@ -270,7 +270,7 @@ class NodeHandler(object):
         # Try to be clever about when to use smartypants
         if node.__class__ in (docutils.nodes.paragraph,
                 docutils.nodes.block_quote, docutils.nodes.title):
-            return smartyPants(text, smarty)
+            return smartypants(text, smarty)
         return text
 
     # End overridable attributes and methods for textdispatch
@@ -286,6 +286,6 @@ class NodeHandler(object):
         except UnicodeDecodeError:
             pass
 
-        text = self.apply_smartypants(text, client.smarty, node)
+        text = self.apply_smartypants(text, client.smartypants_attributes, node)
         node.pdftext = text
         return text