+++ CHANGES.rst
@@ -10,6 +10,7 @@
 * Do not use logging.basicConfig (Issue 509)
 * Removed Tenjin and switched to Jinja2 (Issue #696)
 * Made literal block shrinking work again (Issue #560)
+* Removed copy of smartypants, use PyPI package instead (Issue 694)
 * Fixed bug in token replacement that broke tables in headers/footers (Issue 612)
 * Added: ``:hl_lines:`` code directive allows highlighting of specific lines (issue 623)
 * Added: Extension metadata for Sphinx is now returned in pdfbuilder (issue 640)